requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
asyncpg
python-dotenv
yfinance